swap snv_indel search class inheritance
